Refactor leapYears to make it more readable
This closely resembles the solution given by TOP, but this isn't just a
copy of that. The decision to refactor the function came naturally as I
pored over my code more. I realized that the former solution was
difficult to read unless you already had prior knowledge about moduli
and boolean algebra. Even with that knowledge, it would still take some
time before you understood what you were looking at, unless you already
knew the problem and solution beforehand.

This is one of those problems where the best code isn't necessarily the
fastest or most efficient with memory. This is a problem where the coded
solution can (and probably should) resemble the plain English solution
as much as possible so that just about anyone can read and undertand
exactly what is happening at a glance.